코드 수정

+++ src/main/java/model/service/LikeListManager.java
@@ -1,13 +1,11 @@
 package model.service;
 
 import java.sql.SQLException;
-import java.util.ArrayList;
 import java.util.List;
 
 import model.dao.LikeListDAO;
 import model.dao.PetSitterDAO;
 import model.dto.LikeList;
-import model.dto.PetSitter;
 
 public class LikeListManager {
 	private static LikeListManager likelistMan = new LikeListManager();
@@ -27,52 +25,25 @@ public class LikeListManager {
 		return likelistMan;
 	}
 
-	/* 특정 회원의 likeList에 포함되는 돌보미 정보 반화 */
-	public List<PetSitter> findLikeSitterOfMember(String memberId) throws SQLException {
-		List<LikeList> mLikeLists = likeListDAO.findLikeListOfMember(memberId);
-		List<PetSitter> likeSitterLists = new ArrayList<PetSitter>();
-
-		if (mLikeLists != null) {
-			for (LikeList likeList : mLikeLists) {
-				PetSitter likeSitter = petSitterDAO.findPetSitter(likeList.getLikeSitter().getSitter().getId());
-				String[] address = likeSitter.getSitter().getAddress().split(" ");
-				String city = null;
-				for (int j = 0; j < address.length; j++) {
-					if (address[j].matches("(.*)로")) {
-						city = address[j].substring(0, address[j].length() - 1);
-					}
-				}
-				likeSitter.getSitter().setAddress(city);
-				likeSitterLists.add(likeSitter);
-			}
-		}
-		return likeSitterLists;
-	}
-
 	/* 특정 회원의 likeList 반환 */
 	public List<LikeList> findLikeListOfMember(String memberId) throws SQLException {
 		return likeListDAO.findLikeListOfMember(memberId);
 	}
 
-	/* like 추가 */
-	public boolean add(String memberId, String sitterId) throws SQLException {
-		int count = likeListDAO.add(memberId, sitterId);
+	/* like 추가 또는 삭제 */
+	public boolean addOrRemove(String memberId, String sitterId, String status) throws SQLException {
+		int count = 0;
+		if (status.equals("add"))
+			count = likeListDAO.add(memberId, sitterId);
+		else if (status.equals("remove"))
+			count = likeListDAO.remove(memberId, sitterId);
 		if (count == 0)
 			return false;
-		count = petSitterDAO.updateLikes(sitterId, "add");
-		if (count == 0)
-			return false;
-		return true;
-	}
-	
-	/* like 삭제 */
-	public boolean remove(String memberId, String sitterId) throws SQLException {
-		int count = likeListDAO.remove(memberId, sitterId);
-		if (count == 0)
-			return false;
-		count = petSitterDAO.updateLikes(sitterId, "remove");
+		
+		count = petSitterDAO.updateLikes(sitterId, status);
 		if (count == 0)
 			return false;
 		return true;
 	}
+
 }
